1.0.26: multiple beeps
