+++ AUTHORS
@@ -1,0 +1,5 @@
+https://github.com/Terminus-IMRC/beepALSA
+
+Yukimasa Sugizaki
+James Courtier-Dutton
+Nathan Hurst
