+++ README.md
@@ -1,0 +1,23 @@
+# BEEPALSA
+
+### Description
+
+ALSA implementation `beep`.
+
+### Build
+
+```sh
+make
+```
+
+### Example
+
+```sh
+./beepalsa -f 659 -l 460 -n -f 784 -l 340 -n -f 659 -l 230 -n -f 659 -l 110 -n -f 880 -l 230 -n -f 659 -l 230 -n -f 587 -l 230 -n -f 659 -l 460 -n -f 988 -l 340 -n -f 659 -l 230 -n -f 659 -l 110 -n -f 1047 -l 230 -n -f 988 -l 230 -n -f 784 -l 230 -n -f 659 -l 230 -n -f 988 -l 230 -n -f 1318 -l 230 -n -f 659 -l 110 -n -f 587 -l 230 -n -f 587 -l 110 -n -f 494 -l 230 -n -f 740 -l 230 -n -f 659 -l 460
+```
+
+### Links
+
+https://github.com/Terminus-IMRC/beepALSA
+
+2020
