+++ src/gettext.h
